Run tests for alerts
This is the first iteration of running tests/validations for alerting
rules (and their respective tests, in *_test.yaml).

Pytest is used as a "runner" for simplicity and since it has all
facilities as well. The idea is to integrate with local development, CI
and possibly validation at deploy time too.

A sample of alerts and their tests can be found at Ia0bd232888be

Bug: T272977
Change-Id: I7ae7f8b44fb120af765a676f16765a89ebf216d7