data-protection: add alerting for prometheus-mysqld-exporter failing
Sometimes the prometheus-mysqld-exporter stops exporting (typically
because the mysql/mariadb process has restarted under it); we've done
some work (T289488) to make that less likely, but now additionally add
an alert on the relevant metric so we notice in future if this recurs.

Use "warning", because while it's useful to know about, it's nothing
that requires an immediate response.

Bug: T257056
Change-Id: I89634e4b7af196076329cf180da84602598421b6