Add the first data-engineering team alert to Alertmanager
This is the the first of the Data Engineering team check to be
migrated from Icinga to Alertmanager. It checks the free capacity on the
HDFS system by referring to the Prometheus metrics exported by the
masters.

If the amount of free space is less than 5% it pages the
data-engineering team. If it is less than 10% it sends an email and
alerts the #wikimedia-analytics channel in IRC. If it is less than 15%
it only sends a warning message to the IRC channel.

Paging is restricted to the production cluster.

Unit tests for this check are included.

Bug: T293399
Change-Id: I6daa5997a6a97acd7a3f17053b61afd0609853bb