Use ats-tls metrics for edge traffic drop alert
Due to an ongoing issue with varnishmtail (T293879), VarnishTrafficDrop
is very often reporting false positives. Although eventually the issue
will be fixed, it seems important to (1) rename the alert to
EdgeTrafficDrop, as the software name is an implementation detail, and
(2) use the metrics exposed by ats-tls, which are more reliable than
those currently produced by varnishmtail right now.

Bug: T293879
Change-Id: I89e70ef24c2d054cfbbae5ab77b3df016b7512e8