Add checks for druid datasources to alertmanager
This CR includes two specific alerts:

* Netflow realtime ingestion throughput
* Segments unavailable, per datasource

In a slight deviation from the logic of the Icinga checks, the segments
unavailable check now uses datasource and cluster as dimensions for
aggregation.

This means that we only need to define one alert, but it may fire
multiple times and it will group together the affected datasets.
Similarly, we do not need to define a different alert for the
druid_public cluster.

Bug: T293399
Change-Id: I3c4d2d40b5cffb1eaa0af1e1567ac5c9624ede17