Add CirrusSearch Old GC Hell alerting
Our elasticsearch instances sometimes get stuck in an awkward state
that requires manual attention. Add a new alert that attempts to
recognize when that attention is necessary. When the instances are
healthy the old pool typically shows a sawtooth pattern, when they
are in a problematic state the pool flatlines against the max. Alert
looks for instances that have a mostly flat sawtooth over the last
24 hours. This needs to look over some period of time, as instances in
the alternate cluster show very slow sawtooths that aren't as apparent
in 1 hour snapshots.

Bug: T290604
Change-Id: I76728affc59f7c6c7d8a09c14894145cbc661cbc