Update the way that the unavailable druid segment alert works
This check had been causing some false positive alerts, due to the way
in which Druid defines and then loads the segments. We were
inadvertently alerting on short spikes of unavailable segments, which
were in effect a reasonably small loading queue.

The effect upon whjich we want to alert is if the queue for each
datasource stays above a given threshold value for a significant time.
This commit aims to implement that, so it calculates the average over a
15 minute period for each data source, rather than using the sum of all
measurements taken over a 15 minute period.

Bug: T293399
Change-Id: I9c6f0ea2feef605a1ba4c0422eb7312786fbec24