Adjust PingOffloadMissingIP alert so it recovers properly
The PingOffloadMissingIP alert is designed to catch traffic hitting our
ping offload servers with destination IP addresses not configured on
those hosts.  It is based on the node_netstat_Ip_InAddrErrors metric.

There is a problem with it, however, as that metric is a counter, and
the alert expression will fire if the counter value is >= 1.  Which
means that once the alert fires once it remains active forever.

This change adjusts the expression to use rate(), over the previous 5
minutes, so that it should clear once the counter stops incrementing.

The only part I'm unsure of here is how the unit test needs to be
configured.  Adding rate() there made it fail the Jenkins tests, and
the value bit also confuses me tbh.  Online docs don't explain well.

Bug: T297397
Change-Id: If8cd04da0f3282abf88a9d355458fad5edc9b039