Reduce CirrusSearchJVMGCOldPoolFlatlined false positives
This was alerting based on the high/lo points of a single metric over
time, and has shown to alert in a variety of circumstances that
are some form of flatlining, but not the 'old gc hell' we are trying
to identify. Change the metric to compare the old pool size against
it's maximum allowable size, instead of the maximum size recently
seen.

Longer term we should be using jvmquake to identify these issues
directly.

Change-Id: I8b6ef4d751521062df7062427254efcacc4e61a3