sre: port Kafka alerts from Icinga
The following alerts are a port from check_prometheus
equivalent in modules/profile/manifests/kafka/broker/monitoring.pp:

* KafkaLagIncreasing
* KafkaUnderReplicatedPartitions

Whereas KafkaBrokerUnavailable is meant to replace the check_procs alert
on the kafka broker java process itself. The alert is higher level (i.e.
based on what Kafka is reporting) and should provide good signals.

The intent is to make KafkaBrokerUnavailable a paging alert after a
testing period.

Unlike the Icinga alerts in this case we're aggregating based on the
Kafka cluster and not individual brokers, which AFAIK also provides an
equally good signal.

Bug: T305847
Bug: T309010
Change-Id: Ie642f7dad274ff09410ceb45c0200038262b12f3