elastic: alert on per-node indexing not occurring
WHY THIS CHANGE WAS MADE:

We had an incident where the Elasticsearch threadpool write queue
was increasing monotonically. It turned out the affected instance
was unable to complete any work. This will alert on that condition.

OTHER STUFF:

This might fire when we add a new host to the elasticsearch cluster,
since it will briefly not have any shards.

It will also fire when banning shard allocation on a host once shards
have all moved off of it.

Bug: T314078
Change-Id: I01783586771db6bcb648c059444c44694645ef99