Update the VarnishKafkaNoMessages alert
We have been getting false positive alerts from this alert because it
did not take into account varnish servers that had been intentionally
depooled for maintenance, or even whole datacentres that had been
depooled.

This commit changes the logic of the alert, such that it compares the
rate at which varnish services requests with the rate at which
varnishkafka logs those requests to kafka.

If the rate at which varnishkafka is logging messages is below 20% of
that at which varnish is handling requests, fire a warning. if the rate
is below 10% then fire a critical alert.

Bug: T300246
Change-Id: I12ae4d2c12c70d3594f21b00008550c796b83383