ats: Alert on high connection/request count
Our ATS configuration currently limits the number of active and total
connections with the settings proxy.config.net.max_connections_in and
proxy.config.net.max_connections_active_in respectively. We need to
alert if either of those two values are reached.

Note that the metrics that are monitored are not the ones mentioned
above but custom recording rules. The queries are sufficiently
complex enough that it was worth abstracting them away for simpler
alerting logic.

Additionally, note that trafficserver_active_transaction_count (which
trafficserver_requests_used_percent is based upon) only counts
protocol="http" and not protocol="http2". Servicing HTTP/2 metrics is
tracked in T292817.

Bug: T292815
Change-Id: I04002961f464359792017eebbb58702b0ea18bd4