Update elasticsearch memory pressure alerts
With the upgrade to elasticsearch 7 the behaviour under memory
pressure looks to have changed. Reviewing dashboards it seems like a
better metric will be to monitor how much memory is being used by the
young pool and alert when the young pool is not able to grow as
expected.  Typical healthy instances see hourly young pool maxes
in excess of 1GB, alert when an instance isn't achieving 500MB.

Prompted by elastic1089 OOM'ing while this metric was around 400MB.

Change-Id: I78ceae0d3431f3d5f92db1c2f7c6cc3a4d0e8aee