team-traffic: drop VarnishTrafficDrop and HAProxyEdgeTrafficDrop
Deprecate the VarnishTrafficDrop and HAProxyEdgeTrafficDrop alerts,
which were resulting in alert fatigue.

Further context: after discussion internally in the Traffic team, we
have decided to retire these two alerts. The reason for doing so is
because of the high number of false positives and the resulting alert
fatigue. As an example, these alerts are sent out even when the site is
depooled and in cases of traffic drops resulting from organic changes in
traffic patterns. Pending further work on fixing and improving these
alerts to negate such false positives, we decided that it's best to
retire them and let other metrics such as NEL alert of such changes.

Bug: T322220
Change-Id: I22bdd22fc67f34624d8e000299f85c23ad60248c