KubeletOperationalLatency: Bump to 1s
Increase from 600ms to 1s. This is a big hammer overall, but for the
most part, it should be fine. We 've been seeing alerts flapping lately
for the "list_images" operation. Turns out that once we got enough
images on the Kubernetes nodes filesystem, we have issues listing images
fast enough and this triggers the (admittedly aggressive) p99 alerts.
Bumping the threshold to 1s as one avenue of fixing this for now.
However, given it will happen again soonish, I 'll also dive down to
looking into GC threshold for the kubelet. As much as it is nice to be
able to have tons of OCI images on each node, having 600+ images around
doesn't appear productive.

Change-Id: Ibe96c842b249a6ee8e53fd3bff3cadb2b9bb5785