kafka: update alerts related to brokers down
As far as I can see from the documentation and the metric,
we are monitoring the state (represented as an int number) of
every broker in a cluster. If a broker is running in a healthy
state, its status will be 3+, otherwise less than 3. The alert's
description seems wrong, this change fixes it.

Increase also the alert time from 2m to 5m, since it fires for
every maintenance operation (even a simple restart).

Change-Id: I2f2274b46851077c5dcd2c76cb7190893caf6ce3