search: Add alert based on age of titlesuggest indices
These indices should be updated on a daily basis from a systemd timer.
If for any reason those fail for multiple days in a row on the same
wiki this alert will fire requesting a developer to look into the
error logs.

This approach, querying the update date out of the indices, should be
more reliable than alerting when the build process reports a problem,
any failure to update the relevant indices should trigger this error.

The alert value of 3.25 was chosen with the goal of reducing alerts to
instances where there is almost certainly a problem. This takes into
account that users are unlikely to notice if a single rebuild is
missed, the set of titles that exist on a wiki changes slowly. It also
acknowledges that the current build is not bulletproof and can fail
individual days without there being any concern. The value of 3.25
should avoid triggering the alert if an index arbitrarily fails some
days, but if it fails the same index three days in a row it will alert
and let us know that some larger problem is occurring.

Bug: T327199
Depends-On: If388aaad703ddc4c45b50205c9135efe35c7bde2
Change-Id: I13681c2adc64c90f37cbd0b3f679252ac95c0ea8