Add alert for server-side NIC errors
This adds an alert for server NICs that are reporting errors.

I've left it as a warning in terms of the severity, which I thought
makes sense for introducing it?  But I'm not sure if there is a
better option for that.  Ultimately I think a task to DC-Ops probably
makes sense.

I put the checks in the net_dev file as it seemed to fit with the
existing ones.

Bug: T335350
Change-Id: I70eff1b54e80bff428ee12ca018860d73aef786d