data-engineering: flink: alert based on active site
Application througpuht metrics are valid only when jobs
are consuming kafka topics from the active DC.

For instance, when traffic is routed to eqiad, the codfw Kafka source topic
will won't receive events (save for canary events) and
the in/out ratio will trigger a false positive.
As a workaround, check that events are produced in the input topic on
the Kafka broker beloning to the app deployment site.

Bug: T342258
Bug: T343318
Change-Id: I87ec5d8de9f62c6a2711589d29b8d879782c8e20