team-traffic: add service restart alert for bird
Similar to the changes for haproxy.service, we should be alerted if the
bird service restarts unexpectedly on any of the hosts that are
anycasted, such as internal DNS, Wikimedia DNS and durum.

In theory, we should have individual monitoring for the
anycast-healthchecker service as well but the assumption here is that
the various systemd bindings will take care of that.

Change-Id: I0e5d8468c1c5904332e52128a08077ac455d6404