gitlab: alert on ci failure rate instead of failure ratio
In the initial alert expression we tried to calculate the ratio
between failed ci jobs and all ci jobs. This approach led to false
positive alerts during low usage times (when there is only one active
jobs which fails for example).
The new expression uses the absolute error rate only. The downside is
that we may have to adjust the absolute threshold over time. But for
now this should be a reasonable threshold.

Bug: T339370
Change-Id: If1552d690f51465762f9d9273abe67b805fd1023