team-sre: improve k8s high api latency monitor
Add a new dashboard link for the K8s High latency monitor,
to have a precise idea of what is going on (rather than
seeing a generic dashboard). It should better help in tracking
down a problem when the alert is presented.

I added the latency_percentile variable to the k8s api details
dashboard (set by default to p99) to support this use case
(since it wants p95).

Change-Id: I3ddbd1c3db5f3cd4a7f6a3b8ee1a359ee9408528