Remove host interface errors alert until ethtool stats exposed
This current check is counting errors due to resource exhaustion in
the host operating system network stack, rather than just the kind
of L1 errors it is designed to catch and notfiy dc-ops about.

As detailed in the task we need to move to export ethtool stats
which differentiate between these types of errors, and crerate a
new alert for them when in place.

Until then removing the existing alert as it is causing undue noise
for dc-ops

Bug: T347312
Change-Id: Ifb051f156c1553d51045d5ef7f8c20964a94f834