team-wmcs: ceph: add alert for slow ops
Slow ops usually puts the cluster in warning status, but it has
user-facing impact.

This is not directly paging as there's usually nothing to do except to
wait, but having an alert to point in the right direction makes
troubleshooting easier when paged due to a related user-facing thing
that the slow ops broke.

Change-Id: I3f5f41d1f5dca6a8d2973453923243e6f5a2eca0