team-sre/systemd: update systemd checks to make use of systemd_unit_owner
Users can now set the team parameter on systemd::unit,
systemd::timer::job and systemd::service.  This creates a metric
called systemd_unit_owner allowing useres to override the owner for a
specific unit.

This CR updates the SystemdUnitFailed check to first try
and get the team value from this metric and if that fails fall back to
using the role_owner metric

Bug: T349176
Change-Id: Ie2b7c6dde3c92fbe3df1611dd215c95edc251cd0