team-search-platform: Update job queue alerts to use histogram
Noticed when an alert fired today that the times reported here do not
line up with the current graphs. The graph was reporting ~10m of lag,
but the alert indicated 8h. This variance is curious, but while
investigating I found that the dashboard has switched to using the
histogram_quantile method and a bucket series.

While it's curious that the old method has unclear values, it doesn't
seem worthwhile to figure out why. If general dashboarding has moved
to a different method to calculate this value we should follow suit.
Patch updates our alert to use the same method as the JobQueue Job
dashboard.

Change-Id: I5afe835adf1c2a7bb8424b36ef57242270b54eea