Netfilter: Remove exclude filter.
Keeping track of which hosts are to be alerted on is likely to drift
from Puppet profiles and the available filters may not be sufficient to
ensure that we target the exact hosts.

Removing the filter have no impact on alerting, given that the
utilization calculation comes to zero on hosts without the firewall
profile applied anyway.

Bug: T350694
Change-Id: I7efaaae0cdf94cf707929f9d2d6b7db2600dbb26