PuppetPendingCertificateRequest linting fails in production
The puppet_ca_pending_certificate time series only have data if there is
pending certificate requests. This is due to the instance doing the
requests is added as a label to the time series, leaving us with one
time series per host, with pending requests. Most of the time we have
zero pending requests, hence no data for the
puppet_ca_pending_certificate time serie and pint will fail.

Tell pint to ignore the time series.

Change-Id: I90a6c692d73340464bb865fe5f0a80b52af07eb4