SwiftTooManyMediaUploads: use subtraction instead of increase()
increase() detects a "reset" every time the metric decreases, causing
erroneous large values and an alert. So instead follow the linked
dashboard, using a time offset subtraction.

Also reduce the "for" time to 5m. The nature of the metric is such that
a sufficient spike in uploads lasting for zero seconds (a Dirac delta)
will always exceed the threshold for 1 hour. We may as well send the
alert at the start of the hour rather than waiting until the middle.

Change-Id: Id332ffc793007c31ff4573c2de7bedf9f3c22375