search: Wait for young pool alert to fail for 5 minutes
This alert looks at the maximum value over the last hour. When nodes
shutdown and leave the cluster the last hour eventually means only
a couple minutes of data covering the final moments before shutdown.
The end result is that two or three datapoints at the end can trigger
a meaningless alert.

This alert isn't a fast-action kind of thing anyways. Require the
failure to match for 5 minutes, instead of only 1, to trigger the
alert.

Change-Id: I786133454febfe8a41c5fe7abc6a41fef5698b2e