Move some of the data-engineering alerts to data-platform
We are attempting to split the alerts that are related to data pipelines
and data flow from those related to the data platform and its availability.

This change causes many of the platform related checks to be routed to
the Data Platform SRE team, rather than to the Data Engineering team.

Bug: T346438
Change-Id: I1207797f2386303d9e06256455be71c37cbbe231