cirrus: Port update rate alert from puppet
This was previously alerting on the update rate from cirrus, but
today we took the last step of moving updates from cirrus to SUP.
That rate is now 0. This alert replaces the previous alert and
serves the same purpose, alert if the update rate is well below
what we typically expect.

Change-Id: Ie1229fd26bd970e8f3ddca6fdcb0a2dd8e287e0e