haproxy: Calculate increase of haproxy restarts
The node_systemd_service_restart_total counter will always be non-zero
on a service that's been unexpectedly restarted. This means that the
alert we have for HAProxy restarting will constantly fire on any failure
long after the incident until we restart the service to reset it to
zero. By using increase() the alert will naturally subside without a
restart.

Bug: T362833
Change-Id: I4ed23f47148ac86cfbe7d70f60037e52008707f0