search: Update Cirrus Saneitizer alert
Updated data collection for the saneitizer now reports the
"oldDocument" problem as part of the fixed action. This is indeed
a fix that is being applied, but it's not a problem to alert on.

Exclude the "oldDocument" problem from alerting to bring this
back to how it was working before the data collection was
updated.

Change-Id: Id89c64c7f6c2c71e9b1027933e69f271a168e8c2