Move PyBal failing BGP sessions alert to team-sre and page
This alert is rare enough that when it happens, we should be alerted via
a page and not a simple IRC/email notification that only goes to Traffic
so far. If PyBal is failing to establish BGP sessions, this warrants a
follow-up by an SRE and that can only happen if the alert is paging.

This happened yesterday when there was some work done in ulsfo when the
site was not depooled and the only way we came to know about it was when
we got an email. However, a paging alert (and a quicker one) could have
helped mitigate the damage a bit.

Change-Id: I7a744a599c695dae70d77fec74d9169eedce2ad2