RdfStreamingUpdaterSpaceUsageTooHigh: move alert from search-platform to data-platform
This alert (high disk space usage for RDF streaming updater) is more
appropriate for SREs, so we're moving it away from search-platform
to the data platform SRE group.

I also pointed the link directly to the relevant section, to make it
easier for those who aren't working on the service every day to
understand what's needed.

Bug: T375109
Change-Id: I63ae6f9aca877e9eaa6ca8909a2620daf5fccec7