Add JobQueueLowTrafficConsumerWidespreadHighLatency
Motivation: When a "usually low traffic" job type suddenly becomes high
traffic, it can lead to an isolation failure of sorts, where other truly
low traffic jobs on the shared consumer experience high delay times.
What we saw in T378385 and T379462 are examples of this.

With this in mind, this change introduces an alert to detect widespread
(over 50% of "active" job types) unusually high (over 2m) p95 execution
delay for rules on the shared low traffic consumer.

This is an impact-based alert, which attempts to detect likely cases of
isolation failure in a fairly low-false-positive way. The thresholds,
percentiles, etc. are somewhat arbitrary, but back-testing against the
past couple of weeks of data shows no false positives, and both recent
incidents are identified correctly. That said, we may need to tune this
in the future.

Using *active* jobs in the denominator makes this alert resistant to
cases where jobs other than the "antagonist" are intermittently starved
for throughput. This should avoid having to introduce a "fallback" alert
for unusually high execution rates to complement this.

Bug: T378609
Change-Id: If4a79ce26b8b0e4bbbaaa5c0b7c8c484fe2d626c