team-wmcs: aggregate kernel alerts over 24h
Right now these alerts fire for 30 minutes (because the metric is
measuring "number of errors in the last 30 minutes), then they
auto-resolve. Given some of these errors could happen very infrequently
but we still want to track them, let's keep the alert firing for 24h.

I also increased the alert level to "critical" for "kernel errors" and
"kernel panic" as they seem more concerning than "kernel taints" and
"kernel warnings". They are so rare that we might consider increasing
the priority to "critical" for all of those. Note: a phab task is
automatically created for "critical" alerts but not for "warning" ones.

Bug: T379378
Change-Id: I1a019b7d5b725b58fc589c1f6bed19bf191e2447