Move low-traffic consumer latency alert to critical
With this change, JobQueueLowTrafficRuleProcessingRateTooHigh is moved
to critical to enable IRC notifications.

To enable this:
1. Alert annotations are updated to point to the newly added runbook and
   dedicated low-traffic consumer debugging dashboard.
2. The trigger duration is extended to 6h (was 1h).

Expanding on #2: Part of the rationale for choosing a shorter duration
initially was to ensure the alert works, albeit catching shorter-lived
issues that are not necessarily actionable in practice.

We've now done that [0], so let's make the duration less agressive,
particularly since we've moved the latency-sensitive async upload jobs
out of the shared consumer. The choice of 6h is arbitrary, but is still
a vast improvement in time-to-detection vs. what we have now (someone
notices slowness and opens a task).

[0] This has fired exactly once and resolved within 5m (the antagonist
workload had stopped and the consumer was already recovering). I do have
some ideas for potential improvements we could make to avoid this in the
future, but will pursue those separately.

Bug: T378609
Change-Id: I2b6b6238094ba33ec988275a8b62e553a6ba04d9