WMCS: fix expr in TooManyCloud*Down
Without the extra parentheses, the expr always evaluates to "true".

I also had to modify the eval_time in all tests, because with the
input_series in the test, the alert triggers after 7 mins.

Finally, I had to change the cloudcontrol test where one host was up,
and was not enough to trigger the alert (all hosts must be down).

This is a follow-up to I4aa55d021a0a0a6c74717353430ef9abb8fb05c6

Bug: T381807
Change-Id: I446fbdaaaad4fa2254e39a2c89cba5aa63c47bff