ripeatlas: remove hardcoded measurements
Removing the filter on the measurement label could trigger an alert
for each different Atlas measurement:
* Anchoring Probes Measurement;
* Anchoring Mesh Measurement.

A link to the measurement in the RIPE Atlas Portal is now exported as a label.

Currently, alerts are expressed as a percentage of the total checks,
with thresholds set to 5% for IPv4 checks and 12% for IPv6 checks.
The thresholds are inspired by the initial hardcoded values used also in
Icinga (35 / approx. 700 for ipv4, 90 / approx. 700 ca for ip v6).

Change-Id: Id9baa997fc3c16ded9af3845c8d0909902f6d527
Bug: T370506