mw-on-k8s: align alerts with "pools" of capacity
In advance of the PHP 8.1 migration, the mw-on-k8s alerts were updated
to aggregate by release, so problems isolated to 8.1-specific releases
would not be masked by aggregation together with 7.4. See:
* I34751710c3859c59b0745dac52d8612e72fb6867
* I0f6c536f61cea6e138db8ccaebb7fd49d2964d7d

For alerts that are directly capacity-sensitive - i.e., PHPFPMTooBusy
and MediaWikiLatencyExceeded - aggregating by release is not quite what
we want: it arguably makes more sense to aggregate across releases that
contribute to a common "pool" of capacity backing a service.

This is particularly true for releases like canaries, where their small
size (and rollout strategy) cause transient alerts during deployments.
Conversely, when main and canary are together overloaded, alerting on
both individually is not helpful.

Instead, PHPFPMTooBusy and MediaWikiLatencyExceeded now aggregated by
our mechanism for defining pools of capacity: routed_via.

This is a compromise between the earlier behavior (aggregate across all
releases) and the more recent behavior of aggregating per release, and
better matches our expectations / intent - e.g., alert if either [main,
canary] or [next] is overloaded, but not canary individually.

There's one gotcha here: providing a release "hint" (i.e., where you
should start looking) in the dashboard link. I've taken the approach of
using the routed_via release here, since it will in general be the
largest capacity contributor among the relevant releases.

While we're doing this, also extend the unit tests a bit, particularly
for PHPFPMTooBusy, to make sure we're getting the expected behavior.

Bug: T389224
Change-Id: Idba1001b3eaacadc6fff5d86a79e872904fae35f