Remove Hadoop worker specific disk checks
We had attempted to configure per-disk checks on the Hadoop workers, but this
didn't work because the node_exporter metrics are only collected by the ops
instance of prometheus.

Since they are failing linting for having zero metrics available, we will
remove these checks.

We may replace them with a per-datanode disk space check, rather than a
per-disk check.

Bug: T389466
Bug: T371080
Change-Id: I6afedde1e610aa111e28465834af00906531e03b