Remove the cloudnativepg backup export checks
We had configured AlertManager to send alerts if the systemd service
named backup_s3_postgresql.service had failed on db1208.

However, this check would not have worked because the alert file was
only deployed to the k8s-dse prometheus instance and that instance does
not collect systemd failures.

Since we already collect systemd failures from this host from the
generic systemd checks, we do not need this additional check. Therefore
we can remove it.

Bug: T389466
Bug: T307102
Change-Id: Ia6411a6af61f91ac9e97e172429eb6c41b954470