search: Remove CirrusSearchJVMGCYoungPoolInsufficient alert
We added these alerts a long time ago to catch problems with the
GC, but they haven't given a useful alert in quite some time.
They do fire occasionally, but always as a false positive.
I suspect the improved circuit breakers around memory pressure
have mostly resolved this problem.

We retain the alerts around high frequency of old pool GC, which
should be sufficient to detect problems in the future.

Change-Id: I52ea2484ce18d45facb7ff94ecfa6ff44ab2c250