Netbox: simplify query
The Netbox scripts/reports have been fix and is now running on a
schedule. This allows us to simplify the queries for the alerting and
solely keep track of rate of failures.

The failure rate of a script will continue to increase, until the error
has been fixed, after which it will drop back down to 0 over time and
the alert can be cleared.

Change URI for jobs over view to a page with a more correct
representation of the actual state of the job status.

Bug: T350694
Change-Id: I54b6801316022aa8637a01eeb6199442664ef985