cirrus-streaming-updater: alert when 0 tasks are registered
it happened that flink got started but did not actually run the job.
The metric we use to monitor the health of the job are based on
jobmanager metrics and thus did not capture the problem.
During the downtime flink reported 0 registered task managers, add alert
on this because we should not consider a job stable if it has no
taskmanager.

Change-Id: I8689fdbff8ab1f01c112588702ff5972ea650045