team-data-persistence: Add predictive disk space alerts
In https://phabricator.wikimedia.org/T396200 there was no alert on IRC
on the data-persistence channel.
This change adds predictive disk space monitoring for db* and es*
hosts with warning and critical levels.
"page" severity is not used to first try the alarm without causing
alert spam.
According to [1] the prediction would have alarmed in advance for
the issue in T396200

[1] https://grafana.wikimedia.org/goto/4dA_EnLNR?orgId=1

Change-Id: I9595d38962a35b14afecc883db91d7c3decdb63f