team-traffic: add dnsbox alert for service status mismatch
This alert will help ensure that if we marked a service (recdns, ntp,
authdnsX) as pooled on confctl, that we are also advertising its VIP on
the DNS hosts. This and the related metrics we are exporting gives us
more insight into the state of these (critical) services and their VIPs.

This is a solution for the lack of visibility we saw in T374619.

Bug: T374619
Change-Id: I54b5c8e91478d6532da06e2bc1336cf5849fe3c5