team-traffic: dnsbox: use metrics anycast_healthchecker_service_state
anycast_healthchecker_service_check_ip_assignment is not an accurate
metric since the IP assignment refers to IP assignment in anycast-hc and
does not actually change when the service goes down. Therefore, we
should anycast_healthchecker_service_state, which is 0 for healthy and
anything else for unhealthy. This is helpful to know that the service is
actually unhealthy but pooled in confctl.

Bug: T374619
Change-Id: I497c3c89700c59f985ce897a1711cf1e4367fe6a