team-netops: add checks against Nokia OSPF status
We can only get this data in gnmi for Nokia right now, through their
custom (not OpenConfig) telemetary path.

The alerts do the same thing as the Icinga checks we have for Juniper
devices (the Nokia switches do not support the OSPF SNMP MIB so that
is also not an option for us.)

The idea is simple.  If an interface 'oper-state' is 4 - meaning
'point-to-point', it should hae one neighbor.  Likewise any neighbor
state reported should be 8 (aka 'full') indicating the handshake
completed and things are ok.

Bug: T405558
Change-Id: I41729373bd292824b3fa8cadb1c991f9d1f9f640