search: alert on index failures
Emit an alert when failing bulk index actions because of:
- missing indices
- unknown errors

The threshold is rather strict for now but no problem were encountered
in the last 12h this new metric is in place.
We could relax this is a bit if proven too noisy but we might take this
opportunity to capture problems early.

Bug: T402629
Change-Id: I54135eb8f77bd1b13b3c35e2557ab1ea5dfad787