[eventgate] Split alerts into global and per-site alerts
Eventgate currently alerts on a per-site basis.
This doesn't work out too well if it's taking in events from an active/passive source, as the passive DC eventgate thinks something is wrong.
So we split the alerts into global and local alerts so we can query across aDCs when needed.

Bug: T405952
Change-Id: Ib1706e661aa2798ec43aaa0c806c7756d481bfcf