team-netops: add rule for packet drops in higher-priority queues
First stab at this.  Judging from previous graphs if we have drops
at 2% we are completly saturating the port, 1% may even be an
appropriate threshold.  But we can see how it goes and judge.

Basic concept is to evaluate the ratio of transmitted packets on
an interface - in all queues apart from the "low priority" one - to
the number of packets dropped from those queues.

This scenario can happen if a link max's out, a scenario our
existing 'saturation' rules will catch.  The idea here is that by
ignoring the low-priority queue the alert will only fire if the
saturation is occuring due to what we deem to be the "more
important" traffic.

Eventually this may allow us to reduce the sensitivity of the
saturation alerts, tolerating that situation for longer before
alerting if we are only dropping "low priority" traffic.

Tested a similar expression manually:
https://phabricator.wikimedia.org/P86736

Bug: T384052
Change-Id: I2d49d1806dd2385c46e4cde11f3f6f19c22d6c3a